requirements.txt
asgiref
backports.zoneinfo==0.2.1;python_version<"3.9"
Brotli
cffi
chardet
click
cssselect2
Django
django-extensions
djangorestframework
djangorestframework-simplejwt
fonttools
html5lib
indiapins
mysqlclient
Pillow
pycparser
pydotplus
pydyf
pygraphviz
PyJWT
pyparsing
pyphen
pytz
reportlab
six
sqlparse
tinycss2
typing-extensions
weasyprint
webencodings
zopfli
requests
telesignenterprise
telesign

